LayoutTests/http/tests/websocket/tests/hybi/close-unref-websocket.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Test if Web Socket is closed while handshaking and unreferenced, it should fire close event at most once.");

window.jsTestIsAsync = true;
if (window.layoutTestController)
    layoutTestController.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 0);

var countCloseEvent = 0;

function endTest()
{
    shouldBeGreaterThanOrEqual("1", "countCloseEvent");
    finishJSTest();
};

var ws = new WebSocket("ws://127.0.0.1:8880/websocket/tests/hybi/echo");
ws.onclose = function() {
    countCloseEvent += 1;
};
ws = null;
gc();
setTimeout("endTest()", 100);
</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
